GOOGLE NOW APPLICATION
This application mimics a general News Feed application integrating NewYork Times, Youtube & Google Calendar API.
This application servers as an exercise to display data from an internet server request. A callback was used to retrieve JSON data from the internet and parsed using retrofit library.  A recycler view was used inside a card layout to display the data to a view-holder. Each individual cards can be clicked to open a fragment layout.